#1668 Add a mixin for userIdForDisplay, add userIdForDisplay to component
