+++ dashboard/src/components/users/UserIdForDisplayMixin.vue
@@ -1,0 +1,36 @@
+/*
+Copyright 2020 SkillTree
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+<script>
+  export default {
+    name: 'UserIdForDisplayMixin',
+    methods: {
+      getUserDisplay(props) {
+        const userDisplay = props.userIdForDisplay ? props.userIdForDisplay : props.userId;
+        const { oAuthProviders } = this.$store.getters.config;
+        if (oAuthProviders) {
+          const indexOfDash = userDisplay.lastIndexOf('-');
+          if (indexOfDash > 0) {
+            const provider = userDisplay.substr(indexOfDash + 1);
+            if (oAuthProviders.includes(provider)) {
+              return userDisplay.substr(0, indexOfDash);
+            }
+          }
+        }
+        return userDisplay;
+      },
+    },
+  };
+</script>
