+++ dashboard/src/components/users/UsersTable.vue
@@ -108,9 +108,11 @@ limitations under the License.
   import SkillsBTable from '@/components/utils/table/SkillsBTable';
   import UsersService from './UsersService';
   import DateCell from '../utils/table/DateCell';
+  import UserIdForDisplayMixin from './UserIdForDisplayMixin';
 
   export default {
     name: 'Users',
+    mixins: [UserIdForDisplayMixin],
     components: {
       DateCell,
       SkillsBTable,
@@ -299,20 +301,6 @@ limitations under the License.
         }
         url += '/users';
         return url;
-      },
-      getUserDisplay(props) {
-        const userDisplay = props.userIdForDisplay ? props.userIdForDisplay : props.userId;
-        const { oAuthProviders } = this.$store.getters.config;
-        if (oAuthProviders) {
-          const indexOfDash = userDisplay.lastIndexOf('-');
-          if (indexOfDash > 0) {
-            const provider = userDisplay.substr(indexOfDash + 1);
-            if (oAuthProviders.includes(provider)) {
-              return userDisplay.substr(0, indexOfDash);
-            }
-          }
-        }
-        return userDisplay;
       },
     },
   };
