+++ service/src/main/java/skills/storage/repos/UserEventsRepo.groovy
@@ -452,15 +452,16 @@ interface UserEventsRepo extends CrudRepository<UserEvent, Integer> {
 
     @Nullable
     @Query(value='''
-        SELECT ue.user_id as userId, COUNT(ue.event_time) as count, MAX(ue.event_time) as date
-        FROM user_events ue, (
+        SELECT ue.user_id as userId, ua.user_id_for_display as userIdForDisplay, COUNT(ue.event_time) as count, MAX(ue.event_time) as date
+        FROM user_events ue, user_attrs ua, (
             SELECT user_id, achieved_on, COUNT(id) AS counts FROM user_achievement 
             WHERE skill_ref_id = :skillRefId
             GROUP BY user_id, user_achievement.achieved_on
         ) AS achievements 
         WHERE 
             ue.skill_ref_id = :skillRefId 
-            AND ue.user_id = achievements.user_id 
+            AND ue.user_id = achievements.user_id
+            AND ue.user_id = ua.user_id 
             AND ue.event_time > achievements.achieved_on 
             AND ue.count >= :minEventCountThreshold
         GROUP BY ue.user_id
@@ -469,14 +470,15 @@ interface UserEventsRepo extends CrudRepository<UserEvent, Integer> {
 
     @Nullable
     @Query(value='''
-        SELECT ue.user_id as userId, COUNT(ue.event_time) as count, MAX(ue.event_time) as date
-        FROM user_events ue, (
+        SELECT ue.user_id as userId, ua.user_id_for_display as userIdForDisplay, COUNT(ue.event_time) as count, MAX(ue.event_time) as date
+        FROM user_events ue, user_attrs ua, (
             SELECT user_id, achieved_on, COUNT(id) AS counts FROM user_achievement 
             WHERE skill_ref_id = :skillRefId
             GROUP BY user_id, user_achievement.achieved_on
         ) AS achievements 
         WHERE 
             ue.skill_ref_id = :skillRefId 
+            AND ue.user_id = ua.user_id
             AND ue.user_id = achievements.user_id 
             AND (SELECT MAX(ue2.event_time) from user_events ue2 WHERE ue2.skill_ref_id = :skillRefId AND ue2.user_id = achievements.user_id) < achievements.achieved_on 
         GROUP BY ue.user_id
